dashboard/requirements.txt
pandas
plotly.express
folium
streamlit_folium